+++ interfaces/Portal/src/app/pages/iframe/iframe.module.ts
@@ -3,9 +3,9 @@ import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { IonicModule } from '@ionic/angular';
 import { TranslateModule } from '@ngx-translate/core';
-import { BannerComponent } from '../components/banner/banner.component';
-import { RegistrationProfileComponent } from '../components/registration-profile/registration-profile.component';
-import { SharedModule } from '../shared/shared.module';
+import { BannerComponent } from '../../components/banner/banner.component';
+import { RegistrationProfileComponent } from '../../components/registration-profile/registration-profile.component';
+import { SharedModule } from '../../shared/shared.module';
 import { RecipientPage } from './recipient.page';
 
 @NgModule({
